style(Gateway): Replace getters and setters with a trait
Moved the getters and setters into a trait for reuse
