+++ src/Gateway.php
@@ -11,7 +11,8 @@ use Omnipay\Common\AbstractGateway;
  */
 class Gateway extends AbstractGateway
 {
-    
+    use ParametersTrait;
+
     public function getName()
     {
         return "IfthenPay";
@@ -23,48 +24,8 @@ class Gateway extends AbstractGateway
             'Entidade' => '',
             'Subentidade' => '',
             'ChaveAntiPhishing' => '',
-            'ChaveAntiPhishingCallback' => ''
+            'ChaveAntiPhishingCallback' => '',
         );
-    }
-
-    public function getEntidade()
-    {
-        return $this->getParameter('Entidade');
-    }
-
-    public function setEntidade($value)
-    {
-        return $this->setParameter('Entidade', $value);
-    }
-
-    public function getSubentidade()
-    {
-        return $this->getParameter('Subentidade');
-    }
-
-    public function setSubentidade($value)
-    {
-        return $this->setParameter('Subentidade', $value);
-    }
-
-    public function getChaveAntiPhishing()
-    {
-        return $this->getParameter('ChaveAntiPhishing');
-    }
-
-    public function setChaveAntiPhishing($value)
-    {
-        return $this->setParameter('ChaveAntiPhishing', $value);
-    }
-
-    public function getChaveAntiPhishingCallback()
-    {
-        return $this->getParameter('ChaveAntiPhishingCallback');
-    }
-
-    public function setChaveAntiPhishingCallback($value)
-    {
-        return $this->setParameter('ChaveAntiPhishingCallback', $value);
     }
 
     public function purchase(array $parameters = array())
@@ -76,5 +37,4 @@ class Gateway extends AbstractGateway
     {
         return $this->createRequest('\Omnipay\Ifthenpay\Message\CompletePurchaseRequest', $parameters);
     }
-    
 }
