+++ src/Message/Response.php
@@ -6,7 +6,6 @@ use Omnipay\Common\Message\AbstractResponse;
 
 class Response extends AbstractResponse
 {
-	
     public function isSuccessful()
     {
         return true;
@@ -15,13 +14,14 @@ class Response extends AbstractResponse
 	public function getMessage()
     {
 		$data = $this->getData();
+
         return $data->resposta;
     }
 	
 	public function getTransactionDetails()
     {
         $data = $this->getData();
+
         return $data;
     }
-    
 }
